refactor: split vimApi.ts into two files
