+++ utils/vimApi.ts
@@ -1,5 +1,13 @@
+/**
+ * Partial representation of the CodeMirror Vim API that we use to define motions, commands, etc.
+ *
+ * References:
+ * https://github.com/replit/codemirror-vim/blob/master/src/vim.js
+ * https://libvoyant.ucr.edu/resources/codemirror/doc/manual.html
+ */
+
 import { Editor as CodeMirrorEditor } from "codemirror";
-import { EditorPosition, Editor as ObsidianEditor } from "obsidian";
+import { EditorPosition } from "obsidian";
 
 export type MotionFn = (
   cm: CodeMirrorEditor,
@@ -13,18 +21,6 @@ export type ActionFn = (
   vimState: any
 ) => void;
 
-export type ObsidianActionFn = (
-  obsidianEditor: ObsidianEditor,
-  cm: CodeMirrorEditor,
-  actionArgs: { repeat: number },
-  vimState: any
-) => void;
-
-/**
- * Partial representation of the CodeMirror Vim API that we use to define motions, commands, etc.
- *
- * Reference: https://github.com/replit/codemirror-vim/blob/master/src/vim.js
- */
 export type VimApi = {
   defineMotion: (name: string, fn: MotionFn) => void;
   defineAction: (name: string, fn: ActionFn) => void;
@@ -36,26 +32,3 @@ export type VimApi = {
     extra: { [x: string]: any }
   ) => void;
 };
-
-export function defineObsidianVimMotion(
-  vimObject: VimApi,
-  motionFn: MotionFn,
-  mapping: string
-) {
-  vimObject.defineMotion(motionFn.name, motionFn);
-  vimObject.mapCommand(mapping, "motion", motionFn.name, undefined, {});
-}
-
-export function defineObsidianVimAction(
-  vimObject: VimApi,
-  getActiveObsidianEditor: () => ObsidianEditor,
-  obsidianActionFn: ObsidianActionFn,
-  mapping: string
-) {
-  const actionFn = (cm: CodeMirrorEditor, actionArgs: { repeat: number }, vimState: any) => {
-    const obsidianEditor = getActiveObsidianEditor();
-    obsidianActionFn(obsidianEditor, cm, actionArgs, vimState);
-  }
-  vimObject.defineAction(obsidianActionFn.name, actionFn);
-  vimObject.mapCommand(mapping, "action", obsidianActionFn.name, undefined, {});
-}
